OSS merge

+++ fiftyone/core/dataset.py
@@ -16,7 +16,7 @@ import os
 import random
 import string
 
-from bson import json_util, ObjectId
+from bson import json_util, ObjectId, DBRef
 import cachetools
 from deprecated import deprecated
 import mongoengine.errors as moe
@@ -395,6 +395,10 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
     @property
     def _is_clips(self):
         return self._sample_collection_name.startswith("clips.")
+
+    @property
+    def _is_dynamic_groups(self):
+        return False
 
     @property
     def media_type(self):
@@ -2367,6 +2371,19 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
     def get_group(self, group_id, group_slices=None):
         """Returns a dict containing the samples for the given group ID.
 
+        Examples::
+
+            import fiftyone as fo
+            import fiftyone.zoo as foz
+
+            dataset = foz.load_zoo_dataset("quickstart-groups")
+
+            group_id = dataset.take(1).first().group.id
+            group = dataset.get_group(group_id)
+
+            print(group.keys())
+            # ['left', 'right', 'pcd']
+
         Args:
             group_id: a group ID
             group_slices (None): an optional subset of group slices to load
@@ -3029,11 +3046,12 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
             ops = []
             if issubclass(label_type, fol._LABEL_LIST_FIELDS):
                 array_field = field + "." + label_type._LABEL_LIST_FIELD
+                query = {array_field: {"$exists": True}}
 
                 if view_ids is not None:
                     ops.append(
                         UpdateMany(
-                            {},
+                            query,
                             {
                                 "$pull": {
                                     array_field: {"_id": {"$in": view_ids}}
@@ -3045,14 +3063,15 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
                 if ids is not None:
                     ops.append(
                         UpdateMany(
-                            {}, {"$pull": {array_field: {"_id": {"$in": ids}}}}
+                            query,
+                            {"$pull": {array_field: {"_id": {"$in": ids}}}},
                         )
                     )
 
                 if tags is not None:
                     ops.append(
                         UpdateMany(
-                            {},
+                            query,
                             {
                                 "$pull": {
                                     array_field: {
@@ -3335,7 +3354,7 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
         Returns:
             a list of saved view names
         """
-        return [view_doc.name for view_doc in self._doc.saved_views]
+        return [view_doc.name for view_doc in self._doc.get_saved_views()]
 
     def save_view(
         self,
@@ -3509,6 +3528,9 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
     def delete_saved_views(self):
         """Deletes all saved views from this dataset."""
         for view_doc in self._doc.saved_views:
+            if isinstance(view_doc, DBRef):
+                continue
+
             view_doc.delete()
 
         self._doc.saved_views = []
@@ -3516,7 +3538,7 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
 
     def _delete_saved_view(self, name):
         view_doc = self._get_saved_view_doc(name, pop=True)
-        if view_doc:
+        if not isinstance(view_doc, DBRef):
             view_id = str(view_doc.id)
             view_doc.delete()
         else:
@@ -3530,7 +3552,7 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
         idx = None
         key = "slug" if slug else "name"
 
-        for i, view_doc in enumerate(self._doc.saved_views):
+        for i, view_doc in enumerate(self._doc.get_saved_views()):
             if name == getattr(view_doc, key):
                 idx = i
                 break
@@ -3552,7 +3574,7 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
 
     def _validate_saved_view_name(self, name, skip=None, overwrite=False):
         slug = fou.to_slug(name)
-        for view_doc in self._doc.saved_views:
+        for view_doc in self._doc.get_saved_views():
             if view_doc is skip:
                 continue
 
@@ -5954,7 +5976,7 @@ class Dataset(foc.SampleCollection, metaclass=DatasetSingleton):
         return [{"$match": {"$expr": {"$eq": ["$" + name_field, slice_name]}}}]
 
     def _attach_groups_pipeline(self, group_slices=None):
-        """A pipeline that attaches the reuested group slice(s) for each
+        """A pipeline that attaches the requested group slice(s) for each
         document and stores them in under ``groups.<slice>`` keys.
         """
         if self.group_field is None:
@@ -6618,21 +6640,33 @@ def _do_load_dataset(obj, name):
 
 def _delete_dataset_doc(dataset_doc):
     for view_doc in dataset_doc.saved_views:
+        if isinstance(view_doc, DBRef):
+            continue
+
         view_doc.delete()
 
     for run_doc in dataset_doc.annotation_runs.values():
+        if isinstance(run_doc, DBRef):
+            continue
+
         if run_doc.results is not None:
             run_doc.results.delete()
 
         run_doc.delete()
 
     for run_doc in dataset_doc.brain_methods.values():
+        if isinstance(run_doc, DBRef):
+            continue
+
         if run_doc.results is not None:
             run_doc.results.delete()
 
         run_doc.delete()
 
     for run_doc in dataset_doc.evaluations.values():
+        if isinstance(run_doc, DBRef):
+            continue
+
         if run_doc.results is not None:
             run_doc.results.delete()
 
@@ -6738,7 +6772,7 @@ def _clone_dataset_or_view(dataset_or_view, name, persistent):
         or dataset.has_brain_runs
         or dataset.has_evaluations
     ):
-        _clone_extras(clone_dataset, dataset._doc)
+        _clone_extras(dataset, clone_dataset)
 
     return clone_dataset
 
@@ -7070,11 +7104,12 @@ def _update_no_overwrite(d, dnew):
     d.update({k: v for k, v in dnew.items() if k not in d})
 
 
-def _clone_extras(dst_dataset, src_doc):
+def _clone_extras(src_dataset, dst_dataset):
+    src_doc = src_dataset._doc
     dst_doc = dst_dataset._doc
 
     # Clone saved views
-    for _view_doc in src_doc.saved_views:
+    for _view_doc in src_doc.get_saved_views():
         view_doc = _clone_view_doc(_view_doc)
         view_doc.dataset_id = dst_doc.id
         view_doc.save(upsert=True)
@@ -7082,7 +7117,7 @@ def _clone_extras(dst_dataset, src_doc):
         dst_doc.saved_views.append(view_doc)
 
     # Clone annotation runs
-    for anno_key, _run_doc in src_doc.annotation_runs.items():
+    for anno_key, _run_doc in src_doc.get_annotation_runs().items():
         run_doc = _clone_run(_run_doc)
         run_doc.dataset_id = dst_doc.id
         run_doc.save(upsert=True)
@@ -7090,7 +7125,7 @@ def _clone_extras(dst_dataset, src_doc):
         dst_doc.annotation_runs[anno_key] = run_doc
 
     # Clone brain method runs
-    for brain_key, _run_doc in src_doc.brain_methods.items():
+    for brain_key, _run_doc in src_doc.get_brain_methods().items():
         run_doc = _clone_run(_run_doc)
         run_doc.dataset_id = dst_doc.id
         run_doc.save(upsert=True)
@@ -7098,7 +7133,7 @@ def _clone_extras(dst_dataset, src_doc):
         dst_doc.brain_methods[brain_key] = run_doc
 
     # Clone evaluation runs
-    for eval_key, _run_doc in src_doc.evaluations.items():
+    for eval_key, _run_doc in src_doc.get_evaluations().items():
         run_doc = _clone_run(_run_doc)
         run_doc.dataset_id = dst_doc.id
         run_doc.save(upsert=True)
